refactor(frontend): re-organize component files
