+++ packages/frontend/src/components/alert/Dialog.tsx
@@ -1,21 +1,20 @@
 import React from 'react';
 import { CSSTransition } from 'react-transition-group';
 
-import useScreenType from '../hooks/useScreenType';
+import useScreenType from '../../hooks/useScreenType';
 import {
-  Styled, mergeClassNames, mergeStyles, conditionalClassName, conditionalStyle,
-} from '../utils/style';
+  Styled, mergeClassNames, mergeStyles, conditionalClassName,
+} from '../../utils/style';
 
-import styles from './Dropdown.module.css';
+import styles from './Dialog.module.css';
 
 interface Props {
-  right?: number;
   visible: boolean;
   onClose: () => void;
 }
 
-const Dropdown: React.FC<Styled<Props>> = ({
-  style, className, visible, onClose, children, right = 0,
+const Dialog: React.FC<Styled<Props>> = ({
+  style, className, visible, onClose, children,
 }) => {
   const TIMEOUT = 400;
 
@@ -41,6 +40,9 @@ const Dropdown: React.FC<Styled<Props>> = ({
       in={visible}
       timeout={400}
       classNames={{
+        appear: styles.invisible,
+        appearActive: styles.beingVisible,
+        appearDone: styles.visible,
         enter: styles.invisible,
         enterActive: styles.beingVisible,
         enterDone: styles.visible,
@@ -53,10 +55,10 @@ const Dropdown: React.FC<Styled<Props>> = ({
         role="button"
         tabIndex={0}
         className={mergeClassNames(
-          'absolute w-100vw h-100wh flex flex-col cursor-default',
+          'absolute w-100vw h-100wh flex flex-col z-dialog cursor-default',
           conditionalClassName({
-            desktop: mergeClassNames('items-end justify-start z-dropdown-desktop', styles.desktop),
-            mobile: mergeClassNames('items-center justify-end z-dropdown-mobile', styles.mobile),
+            desktop: mergeClassNames('items-center justify-center', styles.desktop),
+            mobile: mergeClassNames('items-center justify-end', styles.mobile),
           })(screenType),
           styles[`${initialVisible ? 'visible' : 'invisible'}`],
         )}
@@ -73,37 +75,29 @@ const Dropdown: React.FC<Styled<Props>> = ({
       >
         <div
           style={mergeStyles(
-            conditionalStyle({
-              desktop: {
-                marginRight: right,
-                top: 'calc(env(safe-area-inset-top, 0px) + 64px)',
-              },
-            })(screenType),
             {
-              maxHeight: 'calc(100 * var(--wh) - 1rem)',
+              maxHeight: 'calc(90 * var(--wh))',
+              overflow: 'auto',
             },
             style,
           )}
           className={mergeClassNames(
-            'h-fit max-h-5/6 bg-white overflow-hidden',
+            'h-fit max-h-5/6 px-8 py-8 z-dialog-1 bg-white',
             conditionalClassName({
-              desktop: 'w-96 z-dropdown-desktop-1 shadow-dropdown-desktop rounded-b-8 absolute',
-              mobile: 'w-full z-dropdown-mobile-1 shadow-dropdown-mobile rounded-t-12',
+              desktop: 'w-96 shadow-dropdown-desktop rounded-8',
+              mobile: 'w-full shadow-dropdown-mobile rounded-t-12',
               mobileLandscape: 'max-w-sm',
               mobilePortrait: 'max-w-md',
             })(screenType),
             className,
           )}
         >
-          <div className="px-8 py-8 w-full h-full overflow-auto">
-            {children}
-            <div style={{ height: 'env(safe-area-inset-bottom, 0px)' }} />
-          </div>
+          {children}
+          <div style={{ height: 'env(safe-area-inset-bottom, 0px)' }} />
         </div>
       </div>
-
     </CSSTransition>
   );
 };
 
-export default Dropdown;
+export default Dialog;
