+++ packages/frontend/src/components/buttons/AmbientButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { mergeClassNames, mergeStyles, Styled } from '../utils/style';
+import { mergeClassNames, mergeStyles, Styled } from '../../utils/style';
 
 import styles from './AmbientButton.module.css';
