+++ packages/frontend/src/components/layout/Global.tsx
@@ -6,19 +6,19 @@ import {
 } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 
-import classroomState from '../recoil/classroom';
-import dialogState from '../recoil/dialog';
-import dropdownState from '../recoil/dropdown';
-import toastState from '../recoil/toast';
-import { Styled } from '../utils/style';
+import classroomState from '../../recoil/classroom';
+import dialogState from '../../recoil/dialog';
+import dropdownState from '../../recoil/dropdown';
+import toastState from '../../recoil/toast';
+import { Styled } from '../../utils/style';
+import Dialog from '../alert/Dialog';
+import Dropdown from '../alert/Dropdown';
+import ToastDisplay from '../alert/ToastDisplay';
+import YTPlayer from '../youtube/YTPlayer';
+import YTWrapper from '../youtube/YTWrapper';
 
 import Debug from './Debug';
-import Dialog from './Dialog';
-import Dropdown from './Dropdown';
 import ScreenHeightMeasure from './ScreenHeightMeasure';
-import ToastDisplay from './ToastDisplay';
-import YTPlayer from './YTPlayer';
-import YTWrapper from './YTWrapper';
 
 const Global: React.FC<Styled<{}>> = ({ className, style }) => {
   const classroom = useRecoilValue(classroomState.atom);
