+++ packages/frontend/src/components/voice/WaveVisualizer.test.tsx
